Change keyframe interval param to float and add tests

+++ cloudinary-core/src/main/java/com/cloudinary/Transformation.java
@@ -121,7 +121,7 @@ public class Transformation<T extends Transformation> implements Serializable{
         return param("gravity", value);
     }
 
-    public T keyframeInterval(int value) {
+    public T keyframeInterval(float value) {
         return param("keyframe_interval", value);
     }
